Specify version for php-dev
This is a bug I just ran into. If you don't specify the version for `php-dev`, it'll pull dependencies for the LATEST version of PHP, in this case right now, the latest is PHP 7.1, meaning it'll install both PHP 7.0 (as specified for other packages such as `php7.0-cli`) and PHP 7.1 (the latest as per `php-dev`), which is not correct. Simply specifying the version for every single package installed is the solution, and `php-dev` was the only one that didn't. Changing to `php7.0-dev` fixes the issue.